Updated typescript to 2.0-BETA. Renamed configuration-loader to configuration. Fixed typescript errors.
