+++ example/package.json
@@ -14,6 +14,7 @@
     "gulp": "github:gulpjs/gulp#4.0",
     "gulp-sourcemaps": "^1.6.0",
     "gulp-typescript": "^2.13.6",
-    "tsconfig-lint": "^0.11.0"
+    "tsconfig-lint": "^0.11.0",
+    "typescript": "^2.0.0"
   }
 }
