+++ package.json
@@ -13,7 +13,7 @@
     "rollup": "^0.34.1",
     "rollup-watch": "^2.1.1",
     "tslint": "^3.13.0",
-    "typescript": "^1.8.10",
+    "typescript": "^2.0.0",
     "typings": "^1.0.4"
   },
   "dependencies": {
