+++ src/main.ts
@@ -1,5 +1,4 @@
-import * as gulp from 'gulp';
-import Configuration from './configuration/configuration-loader';
+import Configuration from './configuration/configuration';
 import DefaultTasksHandler from './tasks/default/default-tasks-handler';
 
 // Load configuration
