+++ src/paths/paths-builder-base.ts
@@ -3,22 +3,22 @@ import Paths from './paths';
 abstract class BuilderBase {
     protected abstract builder(startPath: string, param: string): string;
 
-    InSource(param: string = undefined) {
+    InSource(param: string) {
         let startPath = Paths.Directories.Source;
         return this.builder(startPath, param);
     }
 
-    InSourceApp(param: string = undefined) {
+    InSourceApp(param: string) {
         let startPath = Paths.Directories.SourceApp;
         return this.builder(startPath, param);
     }
 
-    InBuild(param: string = undefined) {
+    InBuild(param: string) {
         let startPath = Paths.Directories.Build;
         return this.builder(startPath, param);
     }
 
-    InBuildApp(param: string = undefined) {
+    InBuildApp(param: string) {
         let startPath = Paths.Directories.BuildApp;
         return this.builder(startPath, param);
     }
