+++ src/tasks/task-base.ts
@@ -1,9 +1,7 @@
 import { Task } from './tasks-contracts';
-import { TaskFunction } from 'gulp/contracts';
-import Paths from '../paths/paths';
 
 abstract class TaskBase implements Task {
-    Name: string;
+    abstract Name: string;
     public abstract TaskFunction(production: boolean, done?: Function): void | NodeJS.Process | any;
 }
