+++ src/tasks/watch/watch-task.ts
@@ -1,5 +1,4 @@
 import TaskBase from '../task-base';
-import Paths from '../../paths/paths';
 import WatcherTasksHandler from '../../watchers/watcher-tasks-handler';
 
 export default class WatchTask extends TaskBase {
